Initial move to using Knex for talking to database
Initially using it for defining the initial schema + migrations for any changes to the schema.

Also updated the Postgres library to be of a newer version with no native bindings as those wasn’t used anyway and just caused a slow installation process.

Lastly added a new dependency to read a .env file to make the system work without Foreman.