Ignore double slashes in URL:s
Fixes #4

Double slashes are most of the time ignore. Most of the time http://example.com/foo//bar and http://example.com/foo/bar returns the very same page. So whenever a slash is following another slash in the path of a URL they will now be normalized into just a single slash in the normalizing process.